refactor: renamed test files to match method names
