+++ tests/index.js
@@ -3,5 +3,5 @@
 require('./specs/constructor')
 require('./specs/api')
 require('./specs/sync')
-require('./specs/unsynced-local-docs')
-require('./specs/local-changes')
+require('./specs/find-all-unsynced')
+require('./specs/has-local-changes')
